WIP servant 0.12 support experimentation
N.B. this commit has unrelated changes that should be removed.

Alright, so these are some changes I made to try and get it to work
with servant 0.12. Sadly it didn't quite work out as expected. I got
stuck at a point where although everything compiles, type family
reduction gets stuck. The errors that show up are type mismatches
between some type family invocations and the types they evaluate to.
So it looks like a 'mundane stuck types' problem, BUT it does work
in GHCi! This leads me to think that there's a problem in GHC that
only shows up in servant <0.13.

In particular, this commit might make GHC's job a bit easier:

https://github.com/haskell-servant/servant/commit/345dc8a1bddc1ec95adb19475754c9261bbe86c0

So for now the cost/benefit just doesn't work out for me.