Make calls asynchronous
I know it's a bit less pretty now, having to unlift the runner
yourself, but this *must* be the default behavior. Simple as that.

The sync variation might seem useful for a simplistic implementation
of prerendering, but the point of prerendering is reducing latency.
So you will want the concurrency provided by the async variation in
that situation too. Perhaps use something like a DynamicWriterT with
the All Monoid to signal readiness. I don't know about the future,
but right now I don't see any reason to reintroduce the sync
variation.